Move FixedSet to tari_utils
This is a general-purpose struct that is finding use in several places,
so it should live in the general utilities module
